Rename randstr to randchar
As the generator does not return strings but bytes representing characters
