+++ pkg/randchar/fakes/generator.go
